streamlit_app/requirements.txt
streamlit
pandas
psycopg2-binary
sqlalchemy
dotenv